back-django/pyproject.toml
[project]
name = "aviz"
version = "0.1.0"
description = ""
authors = [
    {name = "Aron Sadegh ",email = "aronesadegh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "django-jalali-date (>=1.1.4,<2.0.0)",
    "setuptools (>=78.1.0,<79.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
nplusone = "^1.0.0"

